Move KubernetesGammaServerRunner to separate file
