+++ tools/run_tests/xds_k8s_test_driver/framework/xds_k8s_testcase.py
@@ -44,8 +44,10 @@ from framework.rpc import grpc_csds
 from framework.rpc import grpc_testing
 from framework.test_app import client_app
 from framework.test_app import server_app
+from framework.test_app.runners.k8s import gamma_server_runner
 from framework.test_app.runners.k8s import k8s_xds_client_runner
 from framework.test_app.runners.k8s import k8s_xds_server_runner
+import framework.test_app.runners.k8s.gamma_server_runner
 
 logger = logging.getLogger(__name__)
 # TODO(yashkt): We will no longer need this flag once Core exposes local certs
@@ -66,7 +68,7 @@ TrafficDirectorGammaManager = traffic_director.TrafficDirectorGammaManager
 XdsTestServer = server_app.XdsTestServer
 XdsTestClient = client_app.XdsTestClient
 KubernetesServerRunner = k8s_xds_server_runner.KubernetesServerRunner
-KubernetesGammaServerRunner = k8s_xds_server_runner.KubernetesGammaServerRunner
+GammaServerRunner = gamma_server_runner.GammaServerRunner
 KubernetesClientRunner = k8s_xds_client_runner.KubernetesClientRunner
 _LoadBalancerStatsResponse = grpc_testing.LoadBalancerStatsResponse
 _LoadBalancerAccumulatedStatsResponse = (
@@ -798,10 +800,10 @@ class AppNetXdsKubernetesTestCase(RegularXdsKubernetesTestCase):
 
 class GammaXdsKubernetesTestCase(RegularXdsKubernetesTestCase):
     td: TrafficDirectorGammaManager
-    server_runner: KubernetesGammaServerRunner
+    server_runner: GammaServerRunner
 
-    def initKubernetesServerRunner(self) -> KubernetesGammaServerRunner:
-        return KubernetesGammaServerRunner(
+    def initKubernetesServerRunner(self) -> GammaServerRunner:
+        return GammaServerRunner(
             k8s.KubernetesNamespace(
                 self.k8s_api_manager, self.server_namespace
             ),
